+++ backend/src/main/java/edu/jmi/openatom/server/openatomsystem/dto/RequestPointRedeemItemDTO.java
@@ -12,7 +12,7 @@ public class RequestPointRedeemItemDTO {
   private String description;
 
   @NotNull(message = "兑换所需积分不能为空")
-  private Integer pointCost;
+  private Long pointCost;
 
   private Integer stock;
   private String status;
